feat: CategoryService interface 생성, 기존 CategoryService class는 CategoryServiceImpl로 이름변경
class간 느슨한 결합을 고려하여 CategoryService interface를 작성
